feat(test): replace RequiredProperty test and new Polymorphic test
